feat(WakaTime): add skeleton chart for loading
